// projects/builder/src/lib/builder/models/table-properties.ts
export class TableProperties {
    position: any;
    dgType: any;
    title = 'Column';
    columnSize = 0;
    fieldName = 'Column';
    dgTypeObj: any;
    align = 'left';
    styleSet = {
        fontSize: 2,
        fontName: 'Poppins',
        textalign: 'center',
        color: '#000000'
    }
}